Set up merge queue checks
Motivation:

We would like to have macOS CI run on each PR but suspect that runner
capacity is insufficient to have this run interactively on each commit.
One solution to this is to use merge queues and only run the macOS
checks once the PR is approved for merging.

Modifications:

* Duplicate PR checks (minus soundness) and run them as part of checks on
a merge.
* Remove macOS checks in `main.yml` for now so we don't duplicate work
  on the merges. We might want to rethink the trigger for this on repos
with merge queues to only run on a timer.

Result:

The repo is configured for use with merge queues.